feat: updated error, content, and key handling
app is now able to implement and handle a custom key for data encryption/decryption by the user. response handling has also seen an overhaul and now  returns a response that is validated outside the error listener. finally, changes have been made to the backend to supplement the features mentioned